[WP/AJ] Refactored SelfRegistrationTask to take out a BootstrapService component which does the actual bootstrapping based on the type.
