+++ registry-client/src/main/java/com/tesco/aqueduct/registry/client/BootstrapService.java
@@ -1,0 +1,86 @@
+package com.tesco.aqueduct.registry.client;
+
+import com.tesco.aqueduct.registry.model.BootstrapType;
+import com.tesco.aqueduct.registry.model.Bootstrapable;
+import com.tesco.aqueduct.registry.model.Resetable;
+import io.micronaut.context.annotation.Property;
+import io.micronaut.context.annotation.Value;
+import jakarta.inject.Named;
+import jakarta.inject.Singleton;
+
+import java.time.Duration;
+
+@Singleton
+public class BootstrapService {
+    private final Bootstrapable provider;
+    private final Bootstrapable pipe;
+    private final Bootstrapable controller;
+    private final Resetable corruptionManager;
+    private final long bootstrapDelayMs;
+
+    public BootstrapService(@Named("provider") final Bootstrapable provider,
+                            @Named("pipe") final Bootstrapable pipe,
+                            @Named("controller") final Bootstrapable controller,
+                            @Named("corruptionManager") Resetable corruptionManager,
+                            @Property(name = "registry.http.interval") String retryInterval,
+                            @Value("${pipe.bootstrap.delay:300000}") final int additionalDelay // 5 minutes extra to allow all nodes to reset
+    ) {
+        this.provider = provider;
+        this.pipe = pipe;
+        this.controller = controller;
+        this.corruptionManager = corruptionManager;
+        this.bootstrapDelayMs = Duration.parse("PT" + retryInterval).toMillis() + additionalDelay;
+    }
+
+    public void bootstrap(BootstrapType bootstrapType) throws Exception {
+        switch (bootstrapType) {
+            case PROVIDER:
+                provider.stop();
+                provider.reset();
+                provider.start();
+                break;
+            case PIPE_AND_PROVIDER:
+                provider.stop();
+                provider.reset();
+                pipe.stop();
+                controller.stop();
+                pipe.reset();
+                pipe.start();
+                controller.start();
+                provider.start();
+                break;
+            case PIPE:
+                pipe.stop();
+                controller.stop();
+                pipe.reset();
+                pipe.start();
+                controller.start();
+                break;
+            case PIPE_WITH_DELAY:
+                pipe.stop();
+                controller.stop();
+                pipe.reset();
+                Thread.sleep(bootstrapDelayMs);
+                pipe.start();
+                controller.start();
+                break;
+            case PIPE_AND_PROVIDER_WITH_DELAY:
+                provider.stop();
+                provider.reset();
+                pipe.stop();
+                controller.stop();
+                pipe.reset();
+                Thread.sleep(bootstrapDelayMs);
+                pipe.start();
+                controller.start();
+                provider.start();
+                break;
+            case CORRUPTION_RECOVERY:
+                provider.stop();
+                provider.reset();
+                pipe.stop();
+                corruptionManager.reset();
+                break;
+        }
+    }
+}
